requirements.txt
# trading/requirements.txt

# Core dependencies
qlib>=0.9.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0

# Machine learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Analysis
scipy>=1.10.0

# Utilities
tqdm>=4.65.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
